third_party/blink/web_tests/platform/generic/fast/js/cross-frame-prototype-expected.txt

PASS x.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
